Update solution result to long

+++ Solution.cs
@@ -2,5 +2,5 @@ namespace Aoc2024;
 
 public interface ISolution
 {
-    public int GetSolution();
+    public long GetSolution();
 }
